refactor(chat): split chat module and remove dead code
